# Conflicts:
#	src/main/kotlin/ingsis/tricolor/operations/service/impls/DefaultApiCalls.kt

+++ src/main/kotlin/ingsis/tricolor/operations/controller/ExecutionController.kt
@@ -68,6 +68,6 @@ class ExecutionController(
         exchange: ServerWebExchange,
     ) {
         val correlationId = exchange.getAttribute<String>(CorrelationIdFilter.CORRELATION_ID_KEY) ?: "default-correlation-id"
-        snippetService.updateFormattedLintedSnippet(snippet.id, snippet.content, correlationId)
+        snippetService.updateFormattedLintedSnippet(snippet.id.toLong(), snippet.content, correlationId)
     }
 }
